+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "app/protobuf"]
-	path = app/protobuf
+[submodule "modules/protobuf"]
+	path = modules/protobuf
 	url = https://github.com/tronprotocol/protocol.git
