+++ build.sbt
@@ -5,18 +5,21 @@ organization := "org.tronscan"
 
 version := "latest"
 
-lazy val root = (project in file(".")).enablePlugins(PlayScala)
+lazy val root = (project in file("."))
+  .enablePlugins(PlayScala)
 
 scalaVersion := "2.12.4"
 
-dependencyOverrides += "com.fasterxml.jackson.module" % "jackson-module-scala_2.12" % "2.9.2"
-//dependencyOverrides += "com.fasterxml.jackson.core" % "jackson-databind" % "2.8.8"
-//dependencyOverrides += "com.fasterxml.jackson.module" % "jackson-module-scala_2.11" % "2.8.8"
+dependencyOverrides ++= Seq(
+  "com.fasterxml.jackson.module" % "jackson-module-scala_2.12" % "2.9.2"
+)
 
-libraryDependencies += guice
-libraryDependencies += "org.scalatestplus.play" %% "scalatestplus-play" % "3.1.2" % Test
 
+// Library Dependencies
 libraryDependencies ++= Seq(
+
+  guice,
+
   "com.google.protobuf" % "protobuf-java" % "3.4.0" % "protobuf",
   "com.google.api.grpc" % "googleapis-common-protos" % "0.0.3" % "protobuf",
   "com.thesamet.scalapb" %% "scalapb-runtime" % scalapb.compiler.Version.scalapbVersion % "protobuf",
@@ -31,17 +34,14 @@ libraryDependencies ++= Seq(
   "commons-codec" % "commons-codec" % "1.11",
   "com.madgag.spongycastle" % "core" % "1.53.0.0",
   "com.madgag.spongycastle" % "prov" % "1.53.0.0",
-//  "com.google.guava" % "guava" % "18.0",
   "org.iq80.leveldb" % "leveldb" % "0.10",
   "org.fusesource.leveldbjni" % "leveldbjni-all" % "1.8",
   "org.apache.commons" % "commons-collections4" % "4.0",
   "com.typesafe" % "config" % "1.3.2",
-//  "com.google.code.findbugs" % "jsr305" % "3.0.0",
   "com.cedarsoftware" % "java-util" % "1.8.0",
   "org.apache.commons" % "commons-lang3" % "3.4",
   "org.apache.commons" % "commons-collections4" % "4.0",
   "com.beust" % "jcommander" % "1.72",
-//  "com.github.etaty" %% "rediscala" % "1.8.0",
   "joda-time" % "joda-time" % "2.3",
 
   // Data Access
@@ -80,22 +80,20 @@ libraryDependencies ++= Seq(
   "com.pauldijou" %% "jwt-play-json" % "0.16.0",
   "com.pauldijou" %% "jwt-play" % "0.16.0",
 
-  "org.ocpsoft.prettytime" % "prettytime" % "4.0.1.Final"
+  "org.ocpsoft.prettytime" % "prettytime" % "4.0.1.Final",
+
+  "org.scalatestplus.play" %% "scalatestplus-play" % "3.1.2" % Test,
+  "com.fasterxml.jackson.module" % "jackson-module-scala_2.12" % "2.9.2"
 
 ) ++ grpcDeps ++ akkaDeps ++ circeDependencies
 
+// Disable API Documentation
 sources in (Compile, doc) := Seq.empty
 publishArtifact in (Compile, packageDoc) := false
 
-// Adds additional packages into Twirl
-//TwirlKeys.templateImports += "org.tronscan.controllers._"
-
-// Adds additional packages into conf/routes
-// play.sbt.routes.RoutesKeys.routesImport += "org.tronscan.binders._"
-
-PB.protoSources in Compile := Seq(file("app/protobuf"), file("target/protobuf_external/google/api"))
-
-PB.includePaths in Compile := Seq(file("app/protobuf"), file("target/protobuf_external"))
+//
+PB.protoSources in Compile := Seq(file("modules/protobuf"), file("target/protobuf_external/google/api"))
+PB.includePaths in Compile := Seq(file("modules/protobuf"), file("target/protobuf_external"))
 
 PB.targets in Compile := Seq(
   scalapb.gen() -> (sourceManaged in Compile).value
